organize folders and namespaces. move usage demonstration to TagsCloudVisualizationUsageExample. make TagsCloudVisualization a "library" project
