+++ cs/TagsCloudVisualization/TagsCloudVisualization.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TagsCloudVisualization", "T
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualizationTests", "TagsCloudVisualizationTests\TagsCloudVisualizationTests.csproj", "{E7E9C5DA-47E0-4515-9ECE-7A394E6F5756}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualizationUsageExample", "TagsCloudVisualizationUsageExample\TagsCloudVisualizationUsageExample.csproj", "{3680F905-47DF-4C18-AD2E-E7467CFF8F73}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{E7E9C5DA-47E0-4515-9ECE-7A394E6F5756}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E7E9C5DA-47E0-4515-9ECE-7A394E6F5756}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E7E9C5DA-47E0-4515-9ECE-7A394E6F5756}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3680F905-47DF-4C18-AD2E-E7467CFF8F73}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3680F905-47DF-4C18-AD2E-E7467CFF8F73}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3680F905-47DF-4C18-AD2E-E7467CFF8F73}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3680F905-47DF-4C18-AD2E-E7467CFF8F73}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
