+++ cs/TagsCloudVisualization/TagsCloudVisualization/CircularCloudLayouter.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
+using TagsCloudVisualization.Interfaces;
+using TagsCloudVisualization.Utils;
 
 namespace TagsCloudVisualization
 {
