+++ cs/TagsCloudVisualization/TagsCloudVisualization/Interfaces/IRectangleLayouter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Interfaces
 {
     public interface IRectangleLayouter
     {
