+++ cs/TagsCloudVisualization/TagsCloudVisualization/Utils/PointExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Utils
 {
     public static class PointExtensions
     {
