+++ cs/TagsCloudVisualization/TagsCloudVisualization/Utils/RectangleExtensions.cs
@@ -2,7 +2,7 @@
 using System.Drawing;
 using System.Linq;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Utils
 {
     public static class RectangleExtensions
     {
