+++ cs/TagsCloudVisualization/TagsCloudVisualization/Utils/SpiralPointScatterer.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Utils
 {
     internal class SpiralPointScatterer
     {
