+++ cs/TagsCloudVisualization/TagsCloudVisualizationTests/CircularCloudLayouterTests.cs
@@ -7,6 +7,7 @@ using FluentAssertions;
 using NUnit.Framework;
 using NUnit.Framework.Interfaces;
 using TagsCloudVisualization;
+using TagsCloudVisualization.Visualization;
 
 namespace TagsCloudVisualizationTests
 {
