+++ cs/TagsCloudVisualization/TagsCloudVisualizationTests/RectangleSizeProvider.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualizationTests
 {
     public class RectangleSizeProvider
     {
