+++ cs/TagsCloudVisualization/TagsCloudVisualizationTests/SpiralPointScattererTests.cs
@@ -3,7 +3,7 @@ using System.Drawing;
 using System.Linq;
 using FluentAssertions;
 using NUnit.Framework;
-using TagsCloudVisualization;
+using TagsCloudVisualization.Utils;
 
 namespace TagsCloudVisualizationTests
 {
