+++ cs/TagsCloudVisualization/TagsCloudVisualizationUsageExample/TagsCloudVisualizationUsageExample.csproj
@@ -1,0 +1,20 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <RootNamespace>TagsCloudVisualizationUsageExample</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\TagsCloudVisualizationTests\TagsCloudVisualizationTests.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <None Include="Visualized Examples\100 horizontal word-like rectangles.png" />
+      <None Include="Visualized Examples\100 word-like rectangles.png" />
+    </ItemGroup>
+
+</Project>
